chore: assert trait
